Refactor html
Push path parsing to util and handle title events, loaded, etc.
